+++ src/molecules/LinkTitleImage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import TitleImage from "../atoms/app-title-image";
+import TitleImage from "../atoms/title-logo";
 
 export default function LinkTitleImage() {
   return (
